+++ src/React/Http/Response.php
@@ -17,6 +17,17 @@ class Response extends EventEmitter implements WritableStreamInterface
     public function __construct(ConnectionInterface $conn)
     {
         $this->conn = $conn;
+
+        $that = $this;
+
+        $this->conn->on('error', function ($error) use ($that) {
+            $that->emit('error', array($error, $that));
+            $that->close();
+        });
+
+        $this->conn->on('drain', function () use ($that) {
+            $that->emit('drain');
+        });
     }
 
     public function isWritable()